requirements.txt
requests
numpy
httpx
loguru
opencv-python
fastapi
pillow
uvicorn
hydra-core
omegaconf
scipy
